Removed wxWidgets as submodule. Removed some wxWidgets code. Fixed various warnings. Fixed EditorDebug|x64 config.

+++ MyFramework/SourceCommon/Shaders/ShaderManager.cpp
@@ -316,7 +316,7 @@ bool BaseShader::LoadAndCompile(GLuint premadeprogramhandle)
                 }
 
                 char blendFuncStr[] = "#define BLENDFUNC";
-                int blendFuncStrLen = strlen( blendFuncStr );
+                int blendFuncStrLen = (int)strlen( blendFuncStr );
                 if( i + blendFuncStrLen < m_pFile->GetFileLength() &&
                     strncmp( &buffer[i], blendFuncStr, blendFuncStrLen ) == 0 )
                 {
